Split out HopContext (now FlowTracer) and Breadcrumb from TracerouteEngineImplContext.
Trying to more cleanly define the roles of these classes:
- TEIC provides information about the network needed to run traces.
- FlowTracer does the actual work of tracing the network.
  - FlowTracer still needs some refactoring, but it's getting better.

HopContext was a bad name because *Context implies "just data" -- no
methods other than accessor-like things.
